Use awesome-nav plugin to generate nav hierarchy
By default, you have to explicitly list out all documentation pages in
mkdocs.yml for them to appear in the navigation sidebar.

awesome-nav generates the nav hierarchy based on the directory
hierarchy. By default it globs all the files in each directory in
alphabetical order, and tries to convert directory names to
human-readable names. In a couple of places we need to override this to
sort a particular page first, or to change the generated name for a
subsection: for example, I prefer "Git Basics" to "Git basics".

https://lukasgeiter.github.io/mkdocs-awesome-nav/